+++ packages/nimble-components/src/table/models/table-update-tracker.ts
@@ -132,7 +132,8 @@ export class TableUpdateTracker<
                 changedColumnProperty,
                 'currentFractionalWidth',
                 'currentPixelWidth',
-                'minPixelWidth'
+                'minPixelWidth',
+                'resizingDisabled'
             )
         ) {
             this.track('columnWidths');