+++ packages/nimble-components/src/theme-provider/tests/theme-provider.react.tsx
@@ -1,4 +1,0 @@
-import { ThemeProvider } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleThemeProvider = wrap(ThemeProvider);