+++ packages/storybook/src/nimble/breadcrumb/breadcrumb.stories.ts
@@ -1,0 +1,112 @@
+import type { Meta, StoryObj } from '@storybook/html';
+import { html, repeat } from '@microsoft/fast-element';
+import { breadcrumbItemTag } from '@ni/nimble-components/dist/esm/breadcrumb-item';
+import { breadcrumbTag } from '@ni/nimble-components/dist/esm/breadcrumb';
+import { BreadcrumbAppearance } from '@ni/nimble-components/dist/esm/breadcrumb/types';
+import { createUserSelectedThemeStory } from '../../utilities/storybook';
+
+interface BreadcrumbArgs {
+    options: ItemArgs[];
+    appearance: keyof typeof BreadcrumbAppearance;
+    allowNavigation: boolean;
+}
+
+interface ItemArgs {
+    href?: string;
+    target?: string;
+    label: string;
+}
+
+interface BreadcrumbItemArgs extends ItemArgs {
+    allowNavigation: boolean;
+}
+
+const metadata: Meta<BreadcrumbArgs> = {
+    title: 'Components/Breadcrumb',
+    parameters: {
+        actions: {}
+    }
+};
+
+export default metadata;
+
+export const _standardBreadcrumb: StoryObj<BreadcrumbArgs> = {
+    // prettier-ignore
+    render: createUserSelectedThemeStory(html`
+        <${breadcrumbTag}
+            appearance="${x => BreadcrumbAppearance[x.appearance]}"
+        >
+            ${repeat(x => x.options, html<ItemArgs, BreadcrumbArgs>`
+                <${breadcrumbItemTag}
+                    href="${x => x.href}"
+                    target="${x => x.target}"
+                    @click="${(_x, c) => c.parent.allowNavigation}"
+                >
+                    ${x => x.label}
+                </${breadcrumbItemTag}>
+            `)}
+        </${breadcrumbTag}>
+`),
+    // eslint-disable-next-line storybook/no-redundant-story-name
+    name: 'Standard Breadcrumb',
+    argTypes: {
+        options: {
+            description:
+                'Nest one or more `<nimble-breadcrumb-item />`s inside `<nimble-breadcrumb />`. Each can optionally set `href`, `target`, etc. '
+                + 'With a standard breadcrumb containing multiple items, the last breadcrumb represents the current page (with no `href` specified, '
+                + 'rendering with a bold font).'
+        },
+        appearance: {
+            options: Object.keys(BreadcrumbAppearance),
+            control: { type: 'radio' }
+        }
+    },
+    args: {
+        options: [
+            {
+                href: '#',
+                label: 'Page 1'
+            },
+            {
+                href: '#',
+                label: 'Page 2'
+            },
+            {
+                label: 'Current (No Link)'
+            }
+        ],
+        appearance: 'default',
+        allowNavigation: false
+    }
+};
+
+export const breadcrumbItem: StoryObj<BreadcrumbItemArgs> = {
+    render: createUserSelectedThemeStory(html`
+        <${breadcrumbTag}>
+            <${breadcrumbItemTag}
+                href="${x => x.href}"
+                target="${x => x.target}"
+                @click="${x => x.allowNavigation}"
+            >
+                Breadcrumb Item
+            </${breadcrumbItemTag}>
+        </${breadcrumbTag}>
+    `),
+    argTypes: {
+        href: {
+            description:
+                '(Optional) The URL that this breadcrumb item/ link points to. Generally, the last breadcrumb item '
+                + 'representing the current page has no `href` set.'
+        },
+        target: {
+            description:
+                '(Optional) Where to display the linked URL (destination browsing context): `_self`, `_blank`, etc.',
+            type: 'string'
+        }
+    },
+    args: {
+        href: 'http://www.ni.com',
+        label: 'Breadcrumb Item',
+        allowNavigation: false
+    }
+};