+++ packages/storybook/src/nimble/label-provider/table/label-provider-table.react.tsx
@@ -1,0 +1,4 @@
+import { LabelProviderTable } from '@ni/nimble-components/dist/esm/label-provider/table';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleLabelProviderTable = wrap(LabelProviderTable);