+++ packages/storybook/src/nimble/table-column/text/table-column-text.react.tsx
@@ -1,0 +1,4 @@
+import { TableColumnText } from '@ni/nimble-components/dist/esm/table-column/text';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleTableColumnText = wrap(TableColumnText);