+++ angular-workspace/projects/ni/spright-angular/CHANGELOG.md
@@ -1,8 +1,89 @@
 # Change Log - @ni/spright-angular
 
-This log was last generated on Wed, 24 Apr 2024 18:22:05 GMT and should not be manually modified.
+This log was last generated on Mon, 06 May 2024 19:00:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.11
+
+Mon, 06 May 2024 19:00:05 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.14
+
+## 0.1.10
+
+Thu, 02 May 2024 18:52:31 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.13
+
+## 0.1.9
+
+Wed, 01 May 2024 21:45:37 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.12
+
+## 0.1.8
+
+Wed, 01 May 2024 20:33:43 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.11
+
+## 0.1.7
+
+Wed, 01 May 2024 14:33:42 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.10
+
+## 0.1.6
+
+Tue, 30 Apr 2024 22:26:12 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.9
+
+## 0.1.5
+
+Tue, 30 Apr 2024 21:26:44 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.8
+
+## 0.1.4
+
+Fri, 26 Apr 2024 21:13:31 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.7
+
+## 0.1.3
+
+Fri, 26 Apr 2024 20:31:48 GMT
+
+### Patches
+
+- No API changes, validating release pipelines ([ni/nimble@c5a74e6](https://github.com/ni/nimble/commit/c5a74e68c2909e2f9a50bf0e6b0e4f1c5f0cbce8))
+- Bump @ni/spright-components to v0.0.6
+
+## 0.1.1
+
+Thu, 25 Apr 2024 20:42:08 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.4
 
 ## 0.1.0